requirements.txt
urllib3>=1.15
six>=1.10
certifi
python-dateutil
future
pyyaml
requests
-e git+https://github.com/riseml/config-parser.git@bbc396de7a261ab48875ce37b6277a14fbea654e#egg=config_parser
jsonschema
websocket-client
rollbar
stringcase
-e git+https://github.com/riseml/minio-py.git@add_progress_and_data_streaming_to_uploads#egg=minio
tqdm
pytest-mock